backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
email-validator==2.2.0
python-jose==3.5.0
cryptography==45.0.6
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20